added and tested ChargingStation class

+++ sim_types.h
@@ -7,7 +7,7 @@ class ISimulationAgent
 public:
 	virtual void begin() = 0;
 
-	virtual void timestep_update(unsigned long long prev_time, unsigned long long cur_time) = 0;
+	virtual void timestep_update(size_t prev_time, size_t cur_time) = 0;
 };
